# openapi.yaml
openapi: "3.0.3"
info:
  description: "This is a sample server BIsmarcknserver."
  version: 1.0.0
  title: Bismarc
  contact:
    email: hiroki_hara@hitachi-ch.com
servers:
  - url: http://127.0.0.1:5000
    description: Local enviroment
  - url: https://bismarck-roki3.herokuapp.com/
    description: Cloud enviroment

paths:

  /api/data/{machine_id}:
    post:
      summary: upload raw data file
      description: upload raw data file
      tags: [data]
      deprecated: false

      parameters: 
        - name: machine_id
          in: path
          required: true
          schema: 
            type: string
            example: "ws0001"
      
      requestBody:
        content:
          multipart/form-data:
          #text/csv:
            schema:
              type: object
              properties:
                # 'file' will be the field name in this multipart request

                #machine_id:
                #  type: string
                #  example: "ws0001"
                file:
                  type: string
                  format: binary
                  example: "cash.csv"
                
      responses:
        '201':
          description: Success opertion
          content:
            application/json:
              schema:
                  type: object
                  properties: 
                    result: 
                      type: string
                      example: OK


  /api/models/{machine_id}:
    post:
      summary: execute train
      description: execute train for prediction.
      tags: [models]
      deprecated: false
  
      parameters: 
        - name: machine_id
          in: path
          required: true
          schema: 
            type: string
            example: "ws0001"
  
      responses:
        '200':
          description: Success opertion
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties: 
                    best_model_type:
                      type: string
                      example: reg_mva14
                    currency: 
                      type: string
                      example: 50K

                    mse_test: 
                      type: number
                      format: float
                      example: 86598.94
                    mse_train:
                      type: number
                      format: float
                      example: 65437.42
 
  /api/models/{machine_id}/predict:
    get:
      summary: get predict result
      description: get predict result
      tags: [models]
      deprecated: false

      parameters: 
        - name: machine_id
          in: path
          required: true
          schema: 
            type: string
            example: "ws0001"
        - name: start_date
          in: header
          required: true
          schema: 
            type: string
            example: "2021-11-01"
        - name: end_date
          in: header
          required: true
          schema: 
            type: string
            example: "2021-11-10"

      responses:
        '200':
          description: Success opertion
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties: 
                    currency: 
                      type: string
                      example: 50K
                    predict: 
                      type: array
                      items:
                        type: string
                      example:
                        - "2021-11-01"
                        - -235
                   

  /api/plan/{machine_id}/rep_notes:
    get:
      summary: get optimal number of notes for replenishment
      description: get optimal number of notes for replenishment
      tags: [plan]
      deprecated: false

      parameters: 
        - name: machine_id
          in: path
          required: true
          schema: 
            type: string
            example: "ws0001"
        - name: rep_date
          in: header
          required: true
          schema: 
            type: string
            example: "2021-11-01"

      responses:
        '200':
          description: Success opertion
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties: 
                    currency: 
                      type: string
                      example: 50K
                    incident: 
                      type: number
                      example: 1 
                    max_days: 
                      type: number
                      example: 21                           
                    notes: 
                      type: number
                      example: 2700 

  /api/plan/{machine_id}/incident:
    get:
      summary: get days incident will happen
      description: get days incident will happen
      tags: [plan]
      deprecated: false

      parameters: 
        - name: machine_id
          in: path
          required: true
          schema: 
            type: string
            example: "ws0001"
        - name: pred_start
          in: header
          required: true
          schema: 
            type: string
            example: "2021-11-01"
        - name: cash_position
          in: header
          required: true
          schema: 
            type: number
            example: 1000

      responses:
        '200':
          description: Success opertion
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties: 
                    currency: 
                      type: string
                      example: 50K
                    incident: 
                      type: number
                      example: 1 
                    max_days: 
                      type: number
                      example: 21                           